move kzg and k1 precompile
