+++ crates/precompile/src/secp256k1.rs
@@ -13,11 +13,6 @@
 //!
 //! Output format:
 //! [32 bytes for recovered address]
-#[cfg(feature = "secp256k1")]
-pub mod bitcoin_secp256k1;
-pub mod k256;
-#[cfg(feature = "libsecp256k1")]
-pub mod parity_libsecp256k1;
 
 use crate::{
     utilities::right_pad, PrecompileError, PrecompileOutput, PrecompileResult,
@@ -63,13 +58,6 @@ fn ecrecover_bytes(sig: [u8; 64], recid: u8, msg: [u8; 32]) -> Option<[u8; 32]> 
     }
 }
 
-// Select the correct implementation based on the enabled features.
-cfg_if::cfg_if! {
-    if #[cfg(feature = "secp256k1")] {
-        pub use bitcoin_secp256k1::ecrecover;
-    } else if #[cfg(feature = "libsecp256k1")] {
-        pub use parity_libsecp256k1::ecrecover;
-    } else {
-        pub use k256::ecrecover;
-    }
-}
+// Re-export ecrecover from crypto_provider for backward compatibility
+// TODO: switch to crypto_provider
+pub use crate::crypto_provider::secp256k1::ecrecover;
